+++ packages/core/spec/Rules/Agda_spec.js
@@ -39,7 +39,7 @@ describe('Agda', () => {
       const { rule } = await initialize()
 
       expect(rule.constructCommand()).toEqual({
-        args: ['agda', '--latex', '--latex-dir=.', '$BASE_0'],
+        args: ['agda', '--latex', '--latex-dir=.', '{{$BASE_0}}'],
         cd: '$ROOTDIR/$DIR_0',
         severity: 'error',
         outputs: ['$DIR_0/$NAME_0.tex', '$DIR_0/$NAME_0.agdai', '$DIR_0/agda.sty']