+++ packages/core/spec/Rules/Biber_spec.js
@@ -83,7 +83,7 @@ describe('Biber', () => {
       const { rule } = await initialize()
 
       expect(rule.constructCommand()).toEqual({
-        args: ['biber', '$DIR_0/$BASE_0'],
+        args: ['biber', '{{$FILEPATH_0}}'],
         cd: '$ROOTDIR',
         severity: 'error',
         outputs: ['$DIR_0/$NAME_0.bbl', '$DIR_0/$NAME_0.blg']