+++ packages/core/spec/Rules/MetaPost_spec.js
@@ -52,7 +52,7 @@ describe('MetaPost', () => {
           '-file-line-error',
           '-interaction=batchmode',
           '-recorder',
-          '$BASE_0'
+          '{{$BASE_0}}'
         ],
         cd: '$ROOTDIR_0',
         severity: 'error',