+++ packages/core/spec/Rules/SplitIndex_spec.js
@@ -115,7 +115,7 @@ describe('SplitIndex', () => {
       const { rule } = await initialize()
 
       expect(rule.constructCommand()).toEqual({
-        args: ['splitindex', '-v', '-v', '-m', '', '$DIR_0/$BASE_0'],
+        args: ['splitindex', '-v', '-v', '-m', '', '{{$FILEPATH_0}}'],
         cd: '$ROOTDIR',
         severity: 'error',
         inputs: ['$DIR_0/$NAME_0.log-ParsedSplitIndexStdOut'],