+++ packages/core/src/Rules/MetaPost.js
@@ -22,7 +22,7 @@ export default class MetaPost extends Rule {
         '-file-line-error',
         '-interaction=batchmode',
         '-recorder',
-        '$BASE_0'
+        '{{$BASE_0}}'
       ],
       cd: '$ROOTDIR_0',
       severity: 'error',