+++ packages/core/src/Rules/ParseLaTeXMagic.js
@@ -10,16 +10,17 @@ const ITEM_SEPARATOR_PATTERN = /\s*,\s*/
 export default class ParseLaTeXMagic extends Rule {
   static commands: Set<Command> = new Set(['load'])
   static parameterTypes: Array<Set<string>> = [new Set([
-    'Knitr',
     'LaTeX',
     'LiterateAgda',
-    'LiterateHaskell'
+    'LiterateHaskell',
+    'PythonNoWeb',
+    'RNoWeb'
   ])]
   static defaultActions: Array<Action> = ['parse']
   static description: string = 'Parses Magic comments in LaTeX or knitr documents.'
 
   async parse () {
-    const output = await this.getResolvedOutput('$DIR_0/$BASE_0-ParsedLaTeXMagic')
+    const output = await this.getResolvedOutput('$FILEPATH_0-ParsedLaTeXMagic')
     const magic = {}
 
     await this.firstParameter.parse([{