+++ packages/core/src/Rules/SplitIndex.js
@@ -66,7 +66,7 @@ export default class SplitIndex extends Rule {
   constructCommand (): CommandOptions {
     // Be extra verbose and capture all output for parsing
     return {
-      args: ['splitindex', '-v', '-v', '-m', '', '$DIR_0/$BASE_0'],
+      args: ['splitindex', '-v', '-v', '-m', '', '{{$FILEPATH_0}}'],
       cd: '$ROOTDIR',
       severity: 'error',
       inputs: ['$DIR_0/$NAME_0.log-ParsedSplitIndexStdOut'],