+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Support for PythonTeX \[[#69][]].
+- Support for Pweave \[[#71][]].
+- `knitrOutputPath` option to allow job specific output paths for knitr \[[#71][]].
 
 ## [v0.8.0][] &mdash; 2017-08-15
 
@@ -134,6 +136,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [v0.2.0]: https://github.com/yitzchak/dicy/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/yitzchak/dicy/tree/v0.1.0
 
+[#71]: https://github.com/yitzchak/dicy/pull/71
 [#69]: https://github.com/yitzchak/dicy/pull/69
 [#53]: https://github.com/yitzchak/dicy/pull/53
 [#51]: https://github.com/yitzchak/dicy/pull/51
@@ -143,4 +146,3 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [#27]: https://github.com/yitzchak/dicy/pull/27
 [#23]: https://github.com/yitzchak/dicy/pull/23
 [#21]: https://github.com/yitzchak/dicy/pull/21
-